Use long to count total size and max size of cache, fixes #1

+++ library/src/main/java/im/r_c/android/fusioncache/Cache.java
@@ -70,7 +70,7 @@ interface Cache {
 
     void clear();
 
-    int size();
+    long size();
 
-    int maxSize();
+    long maxSize();
 }
